pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prich"
version = "0.1.0"
description = "A CLI tool for creating and executing reusable LLM prompts with preprocessing and team sharing"
readme = "README.md"
authors = [{name = "Oleksandr Mikriukov", email = "oleks@oleksm.dev"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.7",
    "jinja2>=3.1.4",
    "pyyaml>=6.0.2",
    "rich>=13.7.0",
]
keywords = ["llm", "prompt-engineering", "cli", "team-collaboration", "data-analysis"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
[project.optional-dependencies]
openai = ["openai>=1.0.0"]
mlx = ["mlx_lm>=0.24.1"]
dev = ["pytest", "coverage", "pytest-cov", "twine", "build"]

[project.urls]
Homepage = "https://github.com/oleks-dev/prich"
Repository = "https://github.com/oleks-dev/prich"
Documentation = "https://github.com/oleks-dev/prich#readme"

[project.scripts]
prich = "prich.cli.main:cli"

[project.entry-points."console_scripts"]
prich = "prich.cli.main:cli"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
where = ["."]
include = ["prich*"]
exclude = ["tests*"]

[tool.coverage.run]
branch = true
source = ["prich"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.pytest.ini_options]
addopts = "-ra --cov=prich --cov-report=term-missing"
testpaths = ["tests"]
